Digging in is worth something to the units that need it most
The whole defence calculation is done in fractions -- fortifying is half again,
a river a quarter more, forest and jungle half again -- and every one of those
fractions was thrown away at the end by a cast to a whole number. The attacker's
side of the same sum was never rounded, so the loss was all the defender's.

On the numbers the tests used, tens and hundreds, that cost nothing, which is
why it survived a suite that covers this function closely. On the numbers the
game actually uses it cost the defender the bonus outright: Warriors defend at 1,
so fortifying took them to 1.5 and the cast took them straight back to 1.
Digging in did nothing whatever for the weakest unit in the game -- the one most
often left holding a new city.

That is the shape of "barbarian horsemen kill the fortified city warrior every
time" and "one horse killed two fortified warriors in cities": against a Deity
barbarian's 4.5 the defender was fighting at 1 rather than 1.5, which is the
difference between losing a round four times in five and three times in four.

Deity barbarians are still hard, which is the point of Deity. The dial, if they
want turning down, is the table in DifficultyRules.

The river test asserted the rounded-down answer, so it was pinning the rounding
rather than the rule; it asserts the rule now, and a new test uses a real unit's
defence of 1 rather than a convenient ten.

Refs #61, #75.

+++ Engine/src/Units/UnitExtensions.cs
@@ -42,7 +42,25 @@ public static class UnitExtensions
     }
 
 
-    public static int DefenseFactor(this Unit defendingUnit, Unit attackingUnit, Tile tile, int groundDefMultiplier)
+    /// <summary>
+    /// How hard a unit is to shift, all its bonuses applied.
+    /// </summary>
+    /// <remarks>
+    /// Returned as it is calculated rather than rounded down to a whole number.
+    /// The whole of this is worked out in fractions -- fortifying is half again,
+    /// a river a quarter more, forest and jungle half again -- and every one of
+    /// them used to be thrown away at the end by a cast to int.
+    ///
+    /// On the values the tests used, all tens and hundreds, that cost nothing. On
+    /// the values the game actually uses it cost the defender the bonus outright:
+    /// Warriors defend at 1, so fortifying took them to 1.5 and the cast took them
+    /// straight back to 1. Digging in did nothing whatever for the weakest unit in
+    /// the game, which is the one most often left holding a new city -- and the
+    /// attacker's side of the same sum was never rounded, so the loss was all the
+    /// defender's. It is why a barbarian horseman went through fortified warriors
+    /// in one city after another (#61, #75).
+    /// </remarks>
+    public static double DefenseFactor(this Unit defendingUnit, Unit attackingUnit, Tile tile, int groundDefMultiplier)
     {
         //Carried units cannot be the defender
         if (defendingUnit.InShip != null) return 0;
@@ -181,7 +199,7 @@ public static class UnitExtensions
         // Effect of terrain
         df *= tile.Defense;
 
-        return (int)df;
+        return (double)df;
     }
 
     /// <summary>
